Final changes, some more tidying up and added some extra notes

+++ src/models/reward.py
@@ -1,10 +1,10 @@
-from sqlalchemy import Column, Integer, String, Sequence, BigInteger
+from sqlalchemy import Column, Integer, String, Sequence
 from .base import Base
 
 
 class Reward(Base):
     __tablename__ = 'reward'
-    id = Column(BigInteger, Sequence('reward_id_seq'), primary_key=True)
+    id = Column(Integer, Sequence('reward_id_seq'), primary_key=True)
     name = Column(String())
     category = Column(String())
     points = Column(Integer())
