+++ src/models/user.py
@@ -1,8 +1,8 @@
-from sqlalchemy import Column, BigInteger, Integer
+from sqlalchemy import Column, Integer
 from .base import Base
 
 
 class User(Base):
     __tablename__ = 'user'
-    id = Column(BigInteger, primary_key=True)
+    id = Column(Integer, primary_key=True)
     points = Column(Integer, default=0)
